Fixed problem: download button removes results on click
The problem comes from having the 'Analyse' button which is false as soon as the page reloads. The results are now shown in the dataset01 tab. The table download button is implemented with a button that doesn't reload the page (found on github). This button cannot handle images. The needed extra code is written and works but is not in the button function, yet. Collectri results are missing because of the tab switch.