new: nf-core utility module
The new module works the same ways as the Decoupler utility module. It adds paths and functions to a dataset by inheritance. It allows to create a sample sheet from the path to the raw data and to create the other three needed files for running a nf-core pipline (params file, config file and bash script). At the moment the functionality focuses on the rnaseq nf-core pipeline and running it on the Helix cluster. For running on another cluster, the analysis_params file must be changed as usual. The functionality for running the atacseq pipeline is there, too, but not yet tested since it moved into the standard-workflows package. The added AnalysisI interface makes sure that all datasets officially know about the analysis_params and paths properties.