new: basic metadata handling
Similar to a relational db the metadata should be split up into subject, sample and fastq specific metadata files. fastq specific is placed in the data folder of a dataset. At the moment this is the sample sheet needed as input for nf-core pipelines. The same subj. and sample metadata is used for all datasets of the same sequencing type. On creation of a new analysis object, these two files are merged and saved as tsv.