analysis_params: fixed 'ADD' feature, added 'take_only' feature
added 'take_only' keyword to subset keys after the project_params and dataset_params were merged so that the analysis object knows about all nf-core pipelines used in the project but the single datasets know only the parameters of the pipelines relevant for them. This feature is 'notReady' because of a bugfix in the getpath function. The last element of the path was missing. If FUNKI worked around this bug, the fix might lead to errors. Added test cases for getpath and merge_dicts.